+++ src/components/views/balance/ShowQR.js
