+++ src/components/views/create-wallet/CreateWalletBackupMnemonic.js
