+++ src/components/views/create-wallet/CreateWalletMnemonicConfirm.js
