+++ src/components/views/create-wallet/CreateWalletMnemonicFailed.js
