+++ src/components/views/index.js
@@ -1,22 +1,22 @@
 import Backup from './backup'
-import BackupList from './backup/backup-list'
-import BackupWarning from './backup/backup-warning'
-import BackupTranscribe from './backup/backup-transcribe'
+import BackupList from './backup/BackupList'
+import BackupWarning from './backup/BackupWarning'
+import BackupTranscribe from './backup/BackupTranscribe'
 import Balance from './balance'
 import CreateWallet from './create-wallet'
-import CreateWalletForm from './create-wallet/create-wallet-form'
-import CreateWalletBackupMnemonic from './create-wallet/create-wallet-backup-mnemonic'
-import CreateWalletMnemonicConfirm from './create-wallet/create-wallet-mnemonic-confirm'
-import CreateWalletMnemonicFailed from './create-wallet/create-wallet-mnemonic-failed'
+import CreateWalletForm from './create-wallet/CreateWalletForm'
+import CreateWalletBackupMnemonic from './create-wallet/CreateWalletBackupMnemonic'
+import CreateWalletMnemonicConfirm from './create-wallet/CreateWalletMnemonicConfirm'
+import CreateWalletMnemonicFailed from './create-wallet/CreateWalletMnemonicFailed'
 import Deposit from './deposit'
 import Exit from './exit'
-import ExitForm from './exit/exit-form'
-import ExitConfirm from './exit/exit-confirm'
-import ExitPending from './exit/exit-pending'
+import ExitForm from './exit/ExitForm'
+import ExitConfirm from './exit/ExitConfirm'
+import ExitPending from './exit/ExitPending'
 import History from './history'
 import ImportWallet from './import-wallet'
-import ImportWalletMnemonic from './import-wallet/mnemonic'
-import ImportWalletSuccess from './import-wallet/import-success'
+import ImportWalletMnemonic from './import-wallet/Mnemonic'
+import ImportWalletSuccess from './import-wallet/ImportSuccess'
 import ManageWallet from './manage-wallet'
 import Preview from './preview'
 import Setting from './setting'
